correction ecoulement

+++ inc/Map.class.hpp
@@ -42,6 +42,6 @@ class Map
     	void		elevate_rect(int x0, int y0, int x1, int y1, float value);
         void		draw_cone(int start_x, int start_y, int radius, int height, bool reverse = false);
 	    std::string	to_string(void) const;
-        int			resistance(int x, int y);
-        int			resistance(MapPoint const & point);
+        float		resistance(int x, int y);
+        float		resistance(MapPoint const & point);
 };
